Reorganización en componentes de Dashboard
